\name{Workspace-package}
\alias{Workspace-package}
\alias{Workspace}
\docType{package}
\title{
\packageTitle{Workspace}
}
\description{
\packageDescription{Workspace}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Workspace}
\packageIndices{Workspace}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Workspace}

Maintainer: \packageMaintainer{Workspace}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
